Move type-prepositions.js to utils. Code review
